requirements.txt
opencv-python>=4.9
numpy>=1.26
python-dotenv>=1.0
pytest>=8.0
discord.py>=2.3
tinydb>=4.8